### 2024-05-31 #6820 @editorialbot
[PRE REVIEW]: carbonr: An R Package for Estimating Carbon-Equivalent Emissions
**Submitting author:** @lilyclements (<a href="http://orcid.org/0000-0001-8864-0552">Lily Clements</a>)
**Repository:** https://github.com/IDEMSInternational/carbonr
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
