### 2024-05-31 #6820 @kthyng
@lilyclements
* Do you have docs besides the readme?
* since there is just one author, can you get rid of the "contributed equally" note in the paper?
* the formatting in the paper examples doesn't look very good
* this software looks interesting and useful but the paper doesn't seem to make an argument for being research software, which is the mission for JOSS and a requirement for being in scope. Is this a shift in argument you can make in the paper or would that be a fundamental shift in the software?
