### 2024-06-03 c469d0b
paper updates after JOSS review
