+++ paper.Rmd
@@ -16,7 +16,6 @@ tags:
 authors:
   - name: Lily Clements
   - orcid: 0000-0001-8864-0552
-  - equal-contrib: true
   - affiliation: "1"
 affiliations:
   - name: IDEMS International
@@ -27,73 +26,100 @@ bibliography: paper.bib
 ---
   
 # Summary
-  
 The `carbonr` package provides a user-friendly, open-source tool for calculating carbon-equivalent emissions based on various sources, primarily the UK Government’s greenhouse gas reporting guidelines [@ukgov_greenhouse_gas_reporting_2023].
 
 Designed for flexibility and ease of use, `carbonr` enables users from different sectors to estimate emissions from travel, construction, office environments, and, more recently, clinical settings, thereby aiding in informed decision-making for environmental impact reduction.
 
-# Statement of need
-Climate change and environmental sustainability are pressing global challenges [@WHO2023]. Accurately estimating carbon emissions is important for organisations that aim to understand and mitigate their environmental impact. Despite the significance of this task, there has been a notable lack of open-source tools, which can affect the reproducibility, transparency, and adaptability of existing solutions.
+# Statement of Need
+Climate change and environmental sustainability are pressing global challenges that require sophisticated methods for assessing and mitigating carbon emissions across various sectors [@WHO2023]. Traditional tools for carbon accounting often lack flexibility, transparency, and/or accessibility. These limitations can hamper scientific advancements and the widespread adoption of best practices accounting for in carbon emissions.
 
-The `carbonr` package aims to fill this gap by providing an open-source solution in R, hosted on GitHub. It aims to deliver reliable and reproducible emission estimates, ensuring that results can be reviewed over time. This invites ongoing critique and enhancement of both the code and the estimates, helping `carbonr` to continually improve.
-`carbonr` can offer the flexibility to adjust emission factors and methodologies to suit specific contexts, thereby enhancing the accuracy and relevance of its outputs. By being hosted on GitHub, `carbonr` encourages global community engagement and contributions, enriching the tool with diverse insights and expertise.
+To address these issues, `carbonr` offers an open-source solution that improves the reproducibility of carbon emission calculations and enhances the understanding of these emissions within the research community. Key features of `carbonr` include its simplicity, user-friendly Shiny interface, and adaptability to different sector needs, making it accessible to a broader audience.
 
-The emission calculations are primarily based on the UK Government’s greenhouse gas reporting guidelines [@ukgov_greenhouse_gas_reporting_2023]. Key features of `carbonr` include its simplicity and the integration of a user-friendly Shiny interface, which broadens its accessibility beyond technically skilled users to others interested in estimating their carbon emissions. This approach underscores our commitment to making carbon estimation as straightforward and inclusive as possible.
+By being hosted on GitHub, `carbonr` encourages global community engagement. This collaborative environment not only keeps the tool up-to-date but also allows it to be tailored to specific sectors. A notable enhancement in the application of `carbonr` is its role in assessing emissions from operating theatres, which are significant contributors to hospital carbon footprints. In a collaborative initiative hosted on GitHub, `carbonr` was adapted to include a set of functions that calculate carbon emissions specifically for operating theatres [@ma2024green].
+
+Furthermore, the open-source nature of `carbonr` allows for ongoing verification and comparison of emissions data, this is important for accurate environmental impact assessments.
+
[…]
