### 2024-06-03 2afdbf0
paper table layout update
