+++ paper.md
@@ -65,6 +65,23 @@ library(dplyr)
 airport_finder(name = "Heathrow")
 ```
 
++-------------------+------------+----------+----------+
+| Header 1          | Header 2   | Header 3 | Header 4 |
+|                   |            |          |          |
++:=================:+:==========:+:========:+:========:+
+| row 1, column 1   | column 2   | column 3 | column 4 |
++-------------------+------------+----------+----------+
+| row 2             | cells span columns               |
++-------------------+------------+---------------------+
+| row 3             | cells      | - body              |
++-------------------+ span rows  | - elements          |
+| row 4             |            | - here              |
++===================+============+=====================+
+| Footer                                               |
++===================+============+=====================+
+
+# ...
+
 | Name            | City    | Country        | IATA |
 +:===============:+:=======:+:==============:+:====:+
 | London Heathrow | London  | United Kingdom | LHR  |
