+++ paper.md
@@ -64,21 +64,14 @@ library(dplyr)
 # Finding the airport code for London Heathrow
 airport_finder(name = "Heathrow")
 ```
-
-+-------------------+------------+----------+----------+
-| Header 1          | Header 2   | Header 3 | Header 4 |
-|                   |            |          |          |
-+:=================:+:==========:+:========:+:========:+
-| row 1, column 1   | column 2   | column 3 | column 4 |
-+-------------------+------------+----------+----------+
-| row 2             | cells span columns               |
-+-------------------+------------+---------------------+
-| row 3             | cells      | - body              |
-+-------------------+ span rows  | - elements          |
-| row 4             |            | - here              |
-+===================+============+=====================+
-| Footer                                               |
-+===================+============+=====================+
+| Header 1        | Header 2   | Header 3 | Header 4 |
+|-----------------|------------|----------|----------|
+| row 1, column 1 | column 2   | column 3 | column 4 |
+| row 2           | cells span columns   ||          |
+| row 3           | cells      | - body  |          |
+| row 4           |            | - elements |       |
+|                 |            | - here  |          |
+| **Footer**      |            |          |          |
 
 # ...
 
