### 2024-06-03 #6820 @lilyclements
Hi @kthyng thanks for your feedback and for laying this out so easily to understand. I really appreciate this.

* The documentation is just the readme. The readme is fairly large and so I can quite easily create a vignette to accompany it if the readme is a problem.
* I have now removed the "contributed equally", apologies for that.
* Thanks - I was trying to show the output as it would come from R directly, but I have now changed this to be much more aesthetically appeasing!
* Great point, I didn't intend for it to come across that way. We definitely want to have it relevant to being research software. I have made some shifts in the paper: - namely in "Statement of need" and "Future goals".

I have updated the paper and pushed changes to GitHub.
