### 2024-06-03 #6820 @kthyng
> The readme is fairly large and so I can quite easily create a vignette to accompany it if the readme is a problem.

Agreed that the readme is big and could be repurposed into some other documentation. I would suggest you do this because it would be better for your package but it will probably also come up during the course of your review. You can choose if/when you address it.

Can you take a look at that error?
