+++ paper.md
@@ -16,7 +16,7 @@ tags:
 authors:
   - name: Lily Clements
   - orcid: 0000-0001-8864-0552
-  - affiliation: "1"
+  - affiliation: 1
 affiliations:
   - name: IDEMS International
     index: 1
