+++ vignettes/fragments/overview.Rmd
@@ -1,0 +1,134 @@
+## Overview
+
+carbonr is a package in R to conveniently calculate carbon-equivalent emissions.
+
+The emissions values in the calculations are mostly from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023) when available.
+
+
+## Motivation
+`carbonr` aims to provide a reliable and reproducible approach to calculating emissions levels, ensuring that the results can be saved, edited, and redistributed easily. Further, `carbonr` aims to be transparent in its calculations and values applied. This has the bonus of allowing for flexibility and customisation in estimating carbon-equivalent emissions.
+
+The vision for `carbonr` is to expand and become more comprehensive. We invite contributions from the community to extend the package's functionality, build additional features, and transform it into a more robust tool for estimating carbon-equivalent emissions. Similarly, we invite open discussions and contribution to capture different perspectives and enhancing the functionality of the package.
+
+Finally, `carbonr` aims to make the estimation of carbon-equivalent emissions more accessible by offering a user-friendly front-end interface using Shiny. This ensures that the tools are easier to use, even for individuals with limited programming experience.
+
+## Installation
+
+You can install the development version of carbonr from [GitHub](https://github.com/):
+
+```{r, eval = FALSE}
+# install.packages("devtools")
+devtools::install_github("IDEMSInternational/carbonr")
+```
+
+## Usage
+
+The main function in `naflex` is `na_omit_if`. 
+
+When wrapped around a vector in a summary function, `na_omit_if` ensures that the summary value is calculated when the checks pass, and returns `NA` if not. The example below shows how to calculate the `mean`, conditionally on the proportion of missing values.
+
+```{r}
+library(naflex)
+
+x <- c(1, 3, NA, NA, 3, 2, NA, 5, 8, 7)
+
+# Calculate if 30% or less missing values
+mean(na_omit_if(x, prop = 0.3))
+# Calculate if 20% or less missing values
+mean(na_omit_if(x, prop = 0.2))
+```
+
+Four types of checks are available:
+
+1. `prop`: the maximum proportion (0 to 1) of missing values allowed
+2. `n`: the maximum number of missing values allowed
+3. `consec`: the maximum number of consecutive missing values allowed, and
+4. `n_non`: the minimum number of non-missing values required.
+
+If multiple checks are specified, all checks must pass for missing values to be removed. For example, although there are less than 4 missing values in `x`, there are two consecutive missing values, hence the `consec = 1` check fails in the example below the result is `NA`.
+
+```{r}
+# Calculate if 4 or less missing values and 1 or less consecutive missing values 
+mean(na_omit_if(x, n = 4, consec = 1))
+```
+The use of `%>%` ("pipe") from `magrittr` can be used to make the code look clearer and more familiar. The beginning of the line is now the same as standard R and it moves `na_omit_if` after `x` which then appears more like an option within the function, like `na.rm`, which is how you might think about `na_omit_if` conceptually in this case.
+
+```{r, warning = FALSE}
+require(magrittr)
+
+sum(x %>% na_omit_if(prop = 0.25))
+```
+
+Note that you should not use `na_omit_if` with `na.rm = TRUE` in the summary function since this will always remove missing values so the checks are essentially ignored.
+
+## How `naflex` works & more details
+
+`na_omit_if` works by removing the missing values from `x` if the checks pass, and leaving `x` unmodified otherwise.
+
+```{r}
[…]
