### 2024-06-04 770a678
adding vignette, updating readme
