+++ vignettes/fragments/overview.Rmd
@@ -1,9 +1,8 @@
 ## Overview
 
-carbonr is a package in R to conveniently calculate carbon-equivalent emissions.
+`carbonr` is a package in R to conveniently calculate carbon-equivalent emissions.
 
-The emissions values in the calculations are mostly from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023) when available.
-
+The emissions values in the calculations are mostly from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023) when available. It is specified where the calculations have come from if they are not from the UK Government report (2023).
 
 ## Motivation
 `carbonr` aims to provide a reliable and reproducible approach to calculating emissions levels, ensuring that the results can be saved, edited, and redistributed easily. Further, `carbonr` aims to be transparent in its calculations and values applied. This has the bonus of allowing for flexibility and customisation in estimating carbon-equivalent emissions.
@@ -22,113 +21,242 @@ devtools::install_github("IDEMSInternational/carbonr")
 ```
 
 ## Usage
+Currently in `carbonr`, you can calculate your emissions from raw materials, such as raw fuels, paper, or metal emissions, or you can calculate from different categories such as flights, overall office or household estimated emissions, or vehicles. There are additional options relating specifically to operating theatres as well. Functions include:
 
-The main function in `naflex` is `na_omit_if`. 
+* `airplane_emissions()`
+* `ferry_emissions()`
+* `rail_emissions()`
+* `land_emissions()`
+* `vehicle_emissions()`
+* `hotel_emissions()`
+* `building_emissions()`
+* `office_emissions()`
+* `household_emissions()`
+* `construction_emissions()`
+* `electrical_emissions()`
+* `material_emissions()`
+* `metal_emissions()`
+* `paper_emissions()`
+* `plastic_emissions()`
+* `raw_fuels()`
+* `anaesthetic_emissions()`
+* `clinical_emissions()`
+* `clinical_theatre_data()`
 
-When wrapped around a vector in a summary function, `na_omit_if` ensures that the summary value is calculated when the checks pass, and returns `NA` if not. The example below shows how to calculate the `mean`, conditionally on the proportion of missing values.
+These all return carbon-equivalent emissions in tonnes.
 
-```{r}
-library(naflex)
+A shiny app is also available by `shiny_emissions()` to calculate carbon-equivalent emissions with a GUI.
 
-x <- c(1, 3, NA, NA, 3, 2, NA, 5, 8, 7)
+We give some small examples in using the functions in `carbonr()`.
 
-# Calculate if 30% or less missing values
-mean(na_omit_if(x, prop = 0.3))
-# Calculate if 20% or less missing values
-mean(na_omit_if(x, prop = 0.2))
+```{r, message = FALSE, warning = FALSE}
+library(carbonr)
 ```
 
-Four types of checks are available:
+To calculate emissions for a flight between Vancouver and Toronto, we first want to find the name of the airports. We do this using the `airport_finder()` function:
 
-1. `prop`: the maximum proportion (0 to 1) of missing values allowed
-2. `n`: the maximum number of missing values allowed
-3. `consec`: the maximum number of consecutive missing values allowed, and
-4. `n_non`: the minimum number of non-missing values required.
-
[…]
