+++ README.Rmd
@@ -30,8 +30,7 @@ status](https://www.r-pkg.org/badges/version/carbonr)](https://CRAN.R-project.or
 
 ## Overview
 
-carbonr is a package in R to conveniently calculate carbon-equivalent emissions. The emissions values in the calculations are from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023) whereever available. For more specific functions related to operating theatre waste, alternative sources are used given in the References section.
-Carbon credit prices are additionally available in the `carbon_credit_price` function where values are based on the [World Bank data](https://carbonpricingdashboard.worldbank.org/). The jurisdiction and year available for that jurisdiction can be found in the `check_CPI()` function.
+carbonr is a package in R to conveniently calculate carbon-equivalent emissions. The emissions values in the calculations are from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023) whereever available.
 
 ## Installation
 
@@ -43,7 +42,7 @@ devtools::install_github("IDEMSInternational/carbonr")
 ```
 
 ## Aims of carbonr
-In 2021, work began on the carbonr package in R with the aim of addressing the following challenges and improving the estimation of carbon-equivalent emissions. This came after a review of current approaches to estimate carbon-equivalent emissions. 
+In 2021, work began on the `carbonr` package in R with the aim of addressing the following challenges and improving the estimation of carbon-equivalent emissions. This came after a review of current approaches to estimate carbon-equivalent emissions. 
 
 `carbonr` aims to provide a reliable and reproducible approach to calculating emissions levels, ensuring that the results can be saved, edited, and redistributed easily. Further, `carbonr` aims to be transparent in its calculations and values applied. This has the bonus of allowing for flexibility and customisation in estimating carbon-equivalent emissions.
 
@@ -51,9 +50,9 @@ The vision for `carbonr` is to expand and become more comprehensive. We invite c
 
 Finally, `carbonr` aims to make the estimation of carbon-equivalent emissions more accessible by offering a user-friendly front-end interface using Shiny. This ensures that the tools are easier to use, even for individuals with limited programming experience.
 
-## Functions in carbonr
+## Functions in `carbonr`
 
-carbonr is a package in R to conveniently calculate carbon-equivalent emissions. Currently, emission estimates relate to travel, materials, day-to-day, and clinically based.
+Currently, emission estimates relate to travel, materials, day-to-day, and clinically based.
 
 * `airplane_emissions()`
 * `ferry_emissions()`
@@ -81,7 +80,7 @@ A shiny app is also available by `shiny_emissions()` to calculate carbon-equival
 
 ## Usage
 
-We give some small examples in using the functions in `carbonr()`
+We give some small examples in using the functions in `carbonr()`. Further, more comprehensive, examples are available in the vignette.
 
 ```{r, message = FALSE, warning = FALSE}
 library(carbonr)
@@ -107,54 +106,6 @@ airport_finder(name = "Toronto") %>%
 Now we can find the overall emission value using the appropriate IATA code. These distances are calculated using the Haversine formula:
 ```{r, message = FALSE, warning = FALSE}
 airplane_emissions("YVR", "YTZ")
-```
-
[…]

[…]
