+++ README.md
@@ -20,16 +20,13 @@ status](https://www.r-pkg.org/badges/version/carbonr)](https://CRAN.R-project.or
 
 ## Overview
 
-carbonr is a package in R to conveniently calculate carbon-equivalent
-emissions. The emissions values in the calculations are from the [UK
-Government report
-(2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023)
-whereever available.
+`carbonr` is an R package designed to conveniently calculate
+carbon-equivalent emissions using data from the [UK Government report
+(2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023).
 
 ## Installation
 
-You can install the development version of carbonr from
-[GitHub](https://github.com/) with:
+Install the development version from GitHub with:
 
 ``` r
 # install.packages("devtools")
@@ -38,44 +35,36 @@ devtools::install_github("IDEMSInternational/carbonr")
 
 ## Aims of carbonr
 
-In 2021, work began on the `carbonr` package in R with the aim of
-addressing the following challenges and improving the estimation of
-carbon-equivalent emissions. This came after a review of current
-approaches to estimate carbon-equivalent emissions.
-
-`carbonr` aims to provide a reliable and reproducible approach to
-calculating emissions levels, ensuring that the results can be saved,
-edited, and redistributed easily. Further, `carbonr` aims to be
-transparent in its calculations and values applied. This has the bonus
-of allowing for flexibility and customisation in estimating
-carbon-equivalent emissions.
-
-The vision for `carbonr` is to expand and become more comprehensive. We
-invite contributions from the community to extend the package’s
-functionality, build additional features, and transform it into a more
-robust tool for estimating carbon-equivalent emissions. Similarly, we
-invite open discussions and contribution to capture different
-perspectives and enhancing the functionality of the package.
-
-Finally, `carbonr` aims to make the estimation of carbon-equivalent
-emissions more accessible by offering a user-friendly front-end
-interface using Shiny. This ensures that the tools are easier to use,
-even for individuals with limited programming experience.
+`carbonr` was developed to provide a reliable and reproducible method
+for calculating carbon-equivalent emissions. It aims to: - Ensure
+transparency and flexibility in emissions calculations. - Facilitate
+easy saving, editing, and redistribution of results. - Encourage
+community contributions to extend functionality. - Provide a
+user-friendly interface through Shiny for users with limited programming
+experience.
 
 ## Functions in `carbonr`
 
-Currently, emission estimates relate to travel, materials, day-to-day,
-and clinically based.
+Currently, the package includes functions for estimating emissions from
+various sources:
+
+### Travel-Related Emissions
 
 - `airplane_emissions()`
 - `ferry_emissions()`
 - `rail_emissions()`
 - `land_emissions()`
 - `vehicle_emissions()`
+
+### Accommodation-Related Emissions
+
 - `hotel_emissions()`
 - `building_emissions()`
 - `office_emissions()`
 - `household_emissions()`
+
+### Material-Related Emissions
+
 - `construction_emissions()`
 - `electrical_emissions()`
 - `material_emissions()`
@@ -83,27 +72,28 @@ and clinically based.
 - `paper_emissions()`
 - `plastic_emissions()`
 - `raw_fuels()`
+
+### Clinical Emissions
+
 - `anaesthetic_emissions()`
 - `clinical_emissions()`
 - `clinical_theatre_data()`
 
-These all return carbon-equivalent emissions in tonnes.
-
[…]
