+++ README.Rmd
@@ -51,31 +51,37 @@ devtools::install_github("IDEMSInternational/carbonr")
 
 ## Functions in `carbonr`
 
-Currently, the package includes functions for estimating emissions from various sources:
+#### General Overall Emissions
+These functions estimate emissions based on general activities, making them easier to use without detailed knowledge of specific inputs:
 
-### Travel-Related Emissions
-- `airplane_emissions()`
-- `ferry_emissions()`
-- `rail_emissions()`
-- `land_emissions()`
-- `vehicle_emissions()`
+- **Travel-Related Emissions**
+  - `airplane_emissions()`
+  - `ferry_emissions()`
+  - `rail_emissions()`
+  - `land_emissions()`
+  - `vehicle_emissions()`
 
-### Accommodation-Related Emissions
-- `hotel_emissions()`
-- `building_emissions()`
-- `office_emissions()`
-- `household_emissions()`
+- **Accommodation-Related Emissions**
+  - `hotel_emissions()`
+  - `building_emissions()`
+  - `office_emissions()`
+  - `household_emissions()`
 
-### Material-Related Emissions
-- `construction_emissions()`
-- `electrical_emissions()`
-- `material_emissions()`
-- `metal_emissions()`
-- `paper_emissions()`
-- `plastic_emissions()`
-- `raw_fuels()`
+#### Raw Input-Based Emissions
+These functions allow for more precise calculations by inputting specific values, such as the quantity of materials used:
 
-### Clinical Emissions
+- **Material-Related Emissions**
+  - `construction_emissions()`
+  - `electrical_emissions()`
+  - `material_emissions()`
+  - `metal_emissions()`
+  - `paper_emissions()`
+  - `plastic_emissions()`
+  - `raw_fuels()`
+
+#### Clinical Emissions
+Functions specific to calculating emissions from clinical and medical activities:
+
 - `anaesthetic_emissions()`
 - `clinical_emissions()`
 - `clinical_theatre_data()`
