+++ README.md
@@ -45,35 +45,41 @@ experience.
 
 ## Functions in `carbonr`
 
-Currently, the package includes functions for estimating emissions from
-various sources:
+#### General Overall Emissions
 
-### Travel-Related Emissions
+These functions estimate emissions based on general activities, making
+them easier to use without detailed knowledge of specific inputs:
 
-- `airplane_emissions()`
-- `ferry_emissions()`
-- `rail_emissions()`
-- `land_emissions()`
-- `vehicle_emissions()`
+- **Travel-Related Emissions**
+  - `airplane_emissions()`
+  - `ferry_emissions()`
+  - `rail_emissions()`
+  - `land_emissions()`
+  - `vehicle_emissions()`
+- **Accommodation-Related Emissions**
+  - `hotel_emissions()`
+  - `building_emissions()`
+  - `office_emissions()`
+  - `household_emissions()`
 
-### Accommodation-Related Emissions
+#### Raw Input-Based Emissions
 
-- `hotel_emissions()`
-- `building_emissions()`
-- `office_emissions()`
-- `household_emissions()`
+These functions allow for more precise calculations by inputting
+specific values, such as the quantity of materials used:
 
-### Material-Related Emissions
+- **Material-Related Emissions**
+  - `construction_emissions()`
+  - `electrical_emissions()`
+  - `material_emissions()`
+  - `metal_emissions()`
+  - `paper_emissions()`
+  - `plastic_emissions()`
+  - `raw_fuels()`
 
-- `construction_emissions()`
-- `electrical_emissions()`
-- `material_emissions()`
-- `metal_emissions()`
-- `paper_emissions()`
-- `plastic_emissions()`
-- `raw_fuels()`
+#### Clinical Emissions
 
-### Clinical Emissions
+Functions specific to calculating emissions from clinical and medical
+activities:
 
 - `anaesthetic_emissions()`
 - `clinical_emissions()`
