+++ README.Rmd
@@ -56,48 +56,48 @@ devtools::install_github("IDEMSInternational/carbonr")
 These functions estimate emissions based on general activities, making them easier to use without detailed knowledge of specific inputs:
 
 **Travel-Related Emissions**
-  - `airplane_emissions()`
-  - `ferry_emissions()`
-  - `rail_emissions()`
-  - `land_emissions()`
-  - `vehicle_emissions()`
+ - `airplane_emissions()`
+ - `ferry_emissions()`
+ - `rail_emissions()`
+ - `land_emissions()`
+ - `vehicle_emissions()`
 
 **Accommodation-Related Emissions**
-  - `hotel_emissions()`
-  - `building_emissions()`
-  - `office_emissions()`
-  - `household_emissions()`
+ - `hotel_emissions()`
+ - `building_emissions()`
+ - `office_emissions()`
+ - `household_emissions()`
 
 #### Raw Input-Based Emissions
 These functions allow for more precise calculations by inputting specific values, such as the quantity of materials used:
 
 **Material-Related Emissions**
-  - `construction_emissions()`
-  - `electrical_emissions()`
-  - `material_emissions()`
-  - `metal_emissions()`
-  - `paper_emissions()`
-  - `plastic_emissions()`
-  - `raw_fuels()`
+ - `construction_emissions()`
+ - `electrical_emissions()`
+ - `material_emissions()`
+ - `metal_emissions()`
+ - `paper_emissions()`
+ - `plastic_emissions()`
+ - `raw_fuels()`
 
 #### Clinical Emissions
 Functions specific to calculating emissions from clinical and medical activities:
 
-- `anaesthetic_emissions()`
-- `clinical_emissions()`
-- `clinical_theatre_data()`
+ - `anaesthetic_emissions()`
+ - `clinical_emissions()`
+ - `clinical_theatre_data()`
 
 All functions return carbon-equivalent emissions in tonnes. A Shiny app is available via `shiny_emissions()` for a GUI-based calculation.
 
 ## Usage
 
-Below are examples demonstrating how to use the `carbonr` package functions:
+Below are some small examples demonstrating how to use the `carbonr` package functions. More comprehensive examples can be found in the vignette.
 
 ```{r, message = FALSE, warning = FALSE}
 library(carbonr)
 ```
 
-### Example 1: Calculating Airplane Emissions
+**Calculating Airplane Emissions**
 Find the IATA codes for Vancouver and Toronto:
 
 ```{r, message = FALSE, warning = FALSE, eval=FALSE, include=TRUE}
@@ -121,39 +121,8 @@ Calculate emissions for a flight between these cities:
 airplane_emissions("YVR", "YTZ")
 ```
 
-### Example 2: Emissions for Multiple Individuals
 
-Create a data frame for multiple journeys:
-
-```{r, message = FALSE, warning = FALSE, eval=FALSE, include=TRUE}
-multiple_ind <- tibble::tribble(~ID, ~rail_from, ~rail_to, ~air_from, ~air_to, ~air_via,
-                        "Clint", "Bristol Temple Meads", "Paddington", "LHR", "KIS", "NBO",
-                        "Zara", "Bristol Temple Meads", "Paddington", "LHR", "LAX", "ORL")
-multiple_ind %>%
-  dplyr::rowwise() %>%
-  dplyr::mutate(plane_emissions = airplane_emissions(air_from,
-                                              air_to,
-                                              air_via)) %>%
-  dplyr::mutate(train_emissions = rail_emissions(rail_from,
-                                          rail_to)) %>%
-  dplyr::mutate(total_emissions = plane_emissions + train_emissions)
-```
-```{r, message = FALSE, warning = FALSE, echo = FALSE}
[…]
