+++ vignettes/introduction-to-carbonr.Rmd
@@ -1,11 +1,270 @@
 ---
-title: "Introduction to carbonr"
+title: "Introduction to `carbonr`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{introduction-to-carbonr}
+  %\VignetteIndexEntry{Introduction to `carbonr`}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r child="fragments/overview.Rmd"}
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
 ```
+
+```{r setup}
+library(carbonr)
+```
+
+## Overview
+
+`carbonr` is a package in R to conveniently calculate carbon-equivalent emissions.
+
+The emissions values in the calculations are mostly from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023) when available. It is specified where the calculations have come from if they are not from the UK Government report (2023).
+
+## Motivation
+`carbonr` aims to provide a reliable and reproducible approach to calculating emissions levels, ensuring that the results can be saved, edited, and redistributed easily. Further, `carbonr` aims to be transparent in its calculations and values applied. This has the bonus of allowing for flexibility and customisation in estimating carbon-equivalent emissions.
+
+The vision for `carbonr` is to expand and become more comprehensive. We invite contributions from the community to extend the package's functionality, build additional features, and transform it into a more robust tool for estimating carbon-equivalent emissions. Similarly, we invite open discussions and contribution to capture different perspectives and enhancing the functionality of the package.
+
+Finally, `carbonr` aims to make the estimation of carbon-equivalent emissions more accessible by offering a user-friendly front-end interface using Shiny. This ensures that the tools are easier to use, even for individuals with limited programming experience.
+
+## Installation
+
+You can install the development version of carbonr from [GitHub](https://github.com/):
+
+```{r, eval = FALSE}
+# install.packages("devtools")
+devtools::install_github("IDEMSInternational/carbonr")
+```
+
+## Usage
+Currently in `carbonr`, you can calculate your emissions from raw materials, such as raw fuels, paper, or metal emissions, or you can calculate from different categories such as flights, overall office or household estimated emissions, or vehicles. There are additional options relating specifically to operating theatres as well. Functions include:
+
+* `airplane_emissions()`
+* `ferry_emissions()`
+* `rail_emissions()`
+* `land_emissions()`
+* `vehicle_emissions()`
+* `hotel_emissions()`
+* `building_emissions()`
+* `office_emissions()`
+* `household_emissions()`
+* `construction_emissions()`
+* `electrical_emissions()`
+* `material_emissions()`
+* `metal_emissions()`
+* `paper_emissions()`
+* `plastic_emissions()`
+* `raw_fuels()`
+* `anaesthetic_emissions()`
+* `clinical_emissions()`
+* `clinical_theatre_data()`
+
+These all return carbon-equivalent emissions in tonnes.
+
+A shiny app is also available by `shiny_emissions()` to calculate carbon-equivalent emissions with a GUI.
+
+We give some small examples in using the functions in `carbonr()`.
+
+```{r, message = FALSE, warning = FALSE}
+library(carbonr)
+```
+
[…]
