+++ DESCRIPTION
@@ -10,6 +10,7 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Suggests: 
+    rmarkdown,
     testthat (>= 2.0.0)
 Config/testthat/edition: 2
 Imports: 
@@ -36,3 +37,4 @@ Imports:
 Depends: 
     R (>= 2.10)
 LazyData: true
+VignetteBuilder: knitr
