### 2024-06-04 #6820 @lilyclements
@kthyng I have sorted the error now. It wasn't showing up on my end under Actions so I'm very sorry I didn't flag it sooner.

Thanks for the advice. I have reduced the readme and moved bits into a vignette now. Even if it weren't to come up over the course of the review, I think it makes it a bit more readable now. Thanks.
