### 2024-06-24 #6820 @mengqi-z
👋 @hvwaldow @zhenchun Would any of you be willing to review this submission to JOSS? We conduct our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
