### 2024-06-26 1cfbfb6
Update paper layout for code
