### 2024-06-26 70f1da0
Removing duplicate code
