### 2024-06-26 #6820 @lilyclements
> Hello @lilyclements, I will be the topic editor for your submission. While I am looking for potential reviewers, could you please update the format of the code chunks in your `paper.md` file? For codes in the paper, you don't need to include `{r, message = FALSE, warning = FALSE, eval=FALSE, include=TRUE}` in each code chunk, as it will be printed out in the PDF. Please also ensure the layout of the code in the paper PDF is formatted well. Thank you.

Thanks @mengqi-z, I have done this now.
