### 2024-06-26 #6820 @hvwaldow
> 👋 @hvwaldow @zhenchun Would any of you be willing to review this submission to JOSS? We conduct our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @mengqi-z ! I'll happily review, but will be only able to start Monday next week.
