### 2024-06-26 #6820 @zhenchun
I can review as well.

Zhenchun

On Wed, Jun 26, 2024 at 9:25 AM The Open Journals editorial robot <
***@***.***> wrote:

> *Five most similar historical JOSS papers:*
>
> goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless
> of user experience
> <https://github.com/openjournals/joss-reviews/issues/6393>
> Submitting author: @qepanna
> Handling editor: @arfon (Active)
> Reviewers: @Schuch666, @shubhamjain15, @hahsan1
> Similarity score: 0.7072
>
> The Climate Equity Reference Calculator
> <https://github.com/openjournals/joss-reviews/issues/1273>
> Submitting author: @krueschan
> Handling editor: @jedbrown (Active)
> Reviewers: @mtobis
> Similarity score: 0.7032
>
> bonsai_ipcc - a Python package for the calculation of national greenhouse
> gas inventories <https://github.com/openjournals/joss-reviews/issues/6123>
> Submitting author: @mabudz
> Handling editor: @pdebuyl (Active)
> Reviewers: @GISWLH, @PennyHow
> Similarity score: 0.7032
>
> rfasst: An R tool to estimate air pollution impacts on health and
> agriculture <https://github.com/openjournals/joss-reviews/issues/3820>
> Submitting author: @jonsampedro
> Handling editor: @dhhagan (Retired)
> Reviewers: @nanhung, @ibarraespinosa
> Similarity score: 0.6965
>
> BCEA: An R Package for Cost-Effectiveness Analysis
> <https://github.com/openjournals/joss-reviews/issues/4206>
> Submitting author: @n8thangreen
> Handling editor: @ajstewartlang (Active)
> Reviewers: @rowlandseymour, @MikeLydeamore
> Similarity score: 0.6946
>
> ⚠️ *Note to editors:* If these papers look like they might be a good
> match, click through to the review issue for that paper and invite one or
> more of the authors *before* considering asking the reviewers of these
> papers to review again for JOSS.
>
