### 2024-06-26 #6932 @mengqi-z
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - None
>
> MISSING DOIs
>
> - 10.1163/9789004322714_cclc_2021-0172-549 may be a valid DOI for title: Climate Change and Health
> - No DOI given, and none found for title: Green Theatre Index: Measuring the Effectiveness o...
> - No DOI given, and none found for title: Greenhouse Gas Reporting: Conversion Factors 2023
>
> INVALID DOIs
>
> - None
> ```

@lilyclements Could you please add the missing DOIs in your `paper.bib`? Thank you.
