### 2024-07-03 b1e3a6c
Updating bibliography to include DOIs
