+++ paper.bib
@@ -1,8 +1,9 @@
-@misc{WHO2023,
+@misc{WHO2021,
   author = {{World Health Organisation}},
   title = {Climate Change and Health},
-  year = {2023},
-  url = {https://www.who.int/news-room/fact-sheets/detail/climate-change-and-health}
+  year = {2021},
+  url = {https://www.who.int/news-room/fact-sheets/detail/climate-change-and-health},
+  doi = {10.1163/9789004322714_cclc_2021-0172-549}
 }
 
 @incollection{ma2024green,
@@ -11,7 +12,8 @@
   booktitle={Data-Driven Business Intelligence Systems for Socio-Technical Organizations},
   pages={407--428},
   year={2024},
-  publisher={IGI Global}
+  publisher={IGI Global},
+  doi = {10.4018/979-8-3693-1210-0.ch016}
 }
 
 @misc{ukgov_greenhouse_gas_reporting_2023,
