### 2024-07-03 #6932 @lilyclements
@mengqi-z thanks for this. I have updated to include DOIs in [PR 23](https://github.com/IDEMSInternational/carbonr/pull/23)

For the final one (Greenhouse Gas Reporting), there is not a DOI. Publications, particularly on the GOV.UK site, often do not have DOIs. Is that OK?
