### 2024-07-03 #6932 @mengqi-z
> @mengqi-z thanks for this. I have updated to include DOIs in [PR 23](https://github.com/IDEMSInternational/carbonr/pull/23)
>
> For the final one (Greenhouse Gas Reporting), there is not a DOI. Publications, particularly on the GOV.UK site, often do not have DOIs. Is that OK?

@lilyclements Thank you for fixing the DOIs. If a DOI does not exist for the report, that's fine.

There is one more typo in the `paper.md` file for the citation "Climate Change and Health". According to your `paper.bib`, the citation should be `@WHO2021`, but it is currently cited as `@WHO2023` in `paper.md`. As a result, the rendered PDF does not display the citation correctly.
