### 2024-07-03 ba0475e
Update paper.md
Fixing WHO2021 error
