### 2024-07-03 #6932 @lilyclements
> There is one more typo in the paper.md file for the citation "Climate Change and Health". According to your paper.bib, the citation should be @WHO2021, but it is currently cited as @WHO2023 in paper.md. As a result, the rendered PDF does not display the citation correctly.

@mengqi-z thanks and sorry for that one. I changed that when getting the DOI earlier and forgot to update accordingly. It should render now - thanks!
