### 2024-07-08 IDEMSInternational/carbonr#24 @hvwaldow
JOSS review / formatting of license 
Ref: https://github.com/openjournals/joss-reviews/issues/6932

The file [`LICENSE`](https://github.com/IDEMSInternational/carbonr/blob/main/LICENSE) has weird formatting. It seems Markdown and html elements are used. Here:

https://github.com/IDEMSInternational/carbonr/blob/ba0475ec7ae47af725e39e1d9a1edb7bdbccd91e/LICENSE#L4-L5

and also section headings, lists, ...

This is a pure text - file. Please simply put a copy of the [lgpl-3.0.txt](lgpl-3.0.txt) there.
