### 2024-07-22 IDEMSInternational/carbonr#26 @hvwaldow
JOSS-review : "Via" - parameters in Rail(UK) by shiny_emissions are not accepted by rail_emissions.R
The "Via" - parameters in Rail(UK) by shiny_emissions() are not accepted by rail_emissions.R

if "Via" is set to "1" and "Station" to, e.g., "BPW" then the program fails with

    Error in rail_emissions: Assertion on 'via' failed: Must be of type 'character', not 'list'.
