### 2024-07-22 #6932 @mengqi-z
> > 👋 @zhenchun @hvwaldow - Just checking in to see how the review is going. Could you please provide a brief status update in this thread? There's no rush, but if you anticipate any delays, please let me know. Thanks!
>
> Hi @mengqi-z ! I just finished the first iteration of [my checklist](https://github.com/openjournals/joss-reviews/issues/6932#issuecomment-2214156304). There are four open issues, three very minor and one referring to a bug that needs fixing. I also provided an MR that could be merged to fix another bug. Once these issues have been addressed, I'll have a final look at paper & citations. So that looks pretty good to me 😄

Wonderful! Thank you @hvwaldow !
