### 2024-08-13 bb0b304
adding community guidelines for #28
