### 2024-08-13 8a7f1a8
update paper and bibliography
