+++ paper.md
@@ -32,27 +32,30 @@ The `carbonr` package provides a user-friendly, open-source tool for calculating
 Designed for flexibility and ease of use, `carbonr` enables users from different sectors to estimate emissions from travel, construction, office environments, and, more recently, clinical settings, thereby aiding in informed decision-making for environmental impact reduction.
 
 # Statement of Need
-Climate change and environmental sustainability are pressing global challenges that require sophisticated methods for assessing and mitigating carbon emissions across various sectors [@WHO2023]. Traditional tools for carbon accounting often lack flexibility, transparency, and/or accessibility. These limitations can hamper scientific advancements and the widespread adoption of best practices accounting for in carbon emissions.
+Climate change and environmental sustainability are pressing global challenges that require sophisticated methods for assessing and mitigating carbon emissions across various sectors [@WHO2023]. Accurate carbon accounting is important to understand and reduce environmental impacts, however, the tools available to researchers and practitioners often fall short in terms of flexibility, transparency, and accessibility.
 
-To address these issues, `carbonr` offers an open-source solution that improves the reproducibility of carbon emission calculations and enhances the understanding of these emissions within the research community. Key features of `carbonr` include its simplicity, user-friendly Shiny interface, and adaptability to different sector needs, making it accessible to a broader audience.
+Traditional carbon accounting tools often involve manual calculations using spreadsheets or proprietary software. For example, the UK Government provides Excel-based templates {@ukgov_greenhouse_gas_reporting_2023}, and the Carbon Footprint Calculator is one online calculator to estimate emissions [@carbonfootprint_calculator]. These tools, while comprehensive, are often not easily adaptable to different sectors or use-cases. Additionally, they lack the reproducibility offered by an open-source solution which could automate calculations and provide a more streamlined, adaptable approach. These limitations can hamper scientific advancements and the widespread adoption of best practices in accounting for carbon emissions. 
+
+Current open-source solutions have a limited scope; for example, `co2calculator` in Python focusses on calculations related to heating, electricity, business trips, and transportation [@pledge4future_co2calculator]. Similarly, `footprint` in R specialises in air travel emissions [@rpackage_footprint]. These tools are not comprehensive enough to address the diverse needs of users who require a broader range of emission activities.
+
+To address these issues, `carbonr` offers an open-source solution that improves the reproducibility of carbon emission calculations and enhances the understanding of these emissions within the research community. Key features include its simplicity, user-friendly Shiny interface, and adaptability to different sector needs, making it accessible to a broader audience.
 
 By being hosted on GitHub, `carbonr` encourages global community engagement. This collaborative environment not only keeps the tool up-to-date but also allows it to be tailored to specific sectors. A notable enhancement in the application of `carbonr` is its role in assessing emissions from operating theatres, which are significant contributors to hospital carbon footprints. In a collaborative initiative hosted on GitHub, `carbonr` was adapted to include a set of functions that calculate carbon emissions specifically for operating theatres [@ma2024green].
 
-Furthermore, the open-source nature of `carbonr` allows for ongoing verification and comparison of emissions data, this is important for accurate environmental impact assessments.
+Furthermore, the open-source nature of `carbonr` allows for ongoing verification and comparison of emissions data, this is important for accurate environmental impact assessments. Users can easily modify the tool to suit their specific needs, share their adaptations with the community, and contribute to the continuous improvement of the package.
 
[…]
