+++ paper.bib
@@ -24,3 +24,25 @@
   year = {2023},
   note = {Published on 7 June 2023, last updated on 28 June 2023},
 }
+
+@misc{carbonfootprint_calculator,
+  author = {{Carbon Footprint Ltd.}},
+  title = {Carbon Footprint Calculator},
+  year = {2024},
+  url = {https://www.carbonfootprint.com/calculator.aspx},
+  note = {Accessed: 2024-08-13}
+}
+
+@misc{pledge4future_co2calculator,
+  author = {{Pledge4Future}},
+  title = {CO2 Calculator},
+  year = {2024},
+  url = {https://github.com/pledge4future/co2calculator}
+}
+
+@misc{rpackage_footprint,
+  author = {R Core Team},
+  title = {footprint: Carbon Footprint Calculator for R},
+  year = {2021},
+  url = {https://cran.r-project.org/web/packages/footprint/index.html}
+}
