+++ R/shiny_emissions.R
@@ -211,9 +211,9 @@ shiny_emissions <- function(){
                                                                                                                                                                             width = "41%"),
                                                                                                                                                         shiny::radioButtons(inputId = "typeferry",
                                                                                                                                                                             "Type:",
-                                                                                                                                                                            c("Foot" = "foot",
-                                                                                                                                                                              "Car" = "car",
-                                                                                                                                                                              "Average" = "average")),
+                                                                                                                                                                            c("Foot" = "Foot",
+                                                                                                                                                                              "Car" = "Car",
+                                                                                                                                                                              "Average" = "Average")),
                                                                                                                                                         shiny::numericInput(inputId = "timesferry",
                                                                                                                                                                             label = "Times Journey Made:",
                                                                                                                                                                             value = 1,
