+++ paper.md
@@ -32,7 +32,7 @@ The `carbonr` package provides a user-friendly, open-source tool for calculating
 Designed for flexibility and ease of use, `carbonr` enables users from different sectors to estimate emissions from travel, construction, office environments, and, more recently, clinical settings, thereby aiding in informed decision-making for environmental impact reduction.
 
 # Statement of Need
-Climate change and environmental sustainability are pressing global challenges that require sophisticated methods for assessing and mitigating carbon emissions across various sectors [@WHO2023]. Accurate carbon accounting is important for understanding and reducing environmental impacts; however, the tools available to researchers and practitioners often fall short in terms of flexibility, transparency, and accessibility.
+Climate change and environmental sustainability are pressing global challenges that require sophisticated methods for assessing and mitigating carbon emissions across various sectors [@WHO2021]. Accurate carbon accounting is important for understanding and reducing environmental impacts; however, the tools available to researchers and practitioners often fall short in terms of flexibility, transparency, and accessibility.
 
 Traditional carbon accounting tools often involve manual calculations using spreadsheets or proprietary software. For example, the UK Government provides Excel-based templates {@ukgov_greenhouse_gas_reporting_2023}, and the Carbon Footprint Calculator is one online calculator to estimate emissions [@carbonfootprint_calculator]. These tools, while comprehensive, are often not easily adaptable to different sectors or use-cases. Additionally, they lack the reproducibility offered by an open-source solution which could automate calculations and provide a more streamlined, adaptable approach. These limitations can hamper scientific advancements and the widespread adoption of best practices in accounting for carbon emissions. 
 
