+++ R/clinical_theatre_data.R
@@ -25,7 +25,8 @@
 #' @examples
 #' # Example with dummy data
 #' 
-#' df <- data.frame(time = c("10/04/2000", "10/04/2000", "11/04/2000", "11/04/2000", "12/04/2000", "12/04/2000"),
+#' df <- data.frame(time = c("10/04/2000", "10/04/2000", "11/04/2000",
+#'                           "11/04/2000", "12/04/2000", "12/04/2000"),
 #' theatre = rep(c("A", "B"), times = 3),
 #' clinical_waste = c(80, 90, 80, 100, 120, 110),
 #' electricity_kwh = c(100, 110, 90, 100, 100, 110),
@@ -42,7 +43,8 @@
 #'                  jurisdiction = "Australia",
 #'                  year = 2023,
 #'                  single_sheet = FALSE)
-# x <- clinical_theatre_data(clincial_example_df, time = date_yyyy_mm, name = theatre_name,
+# x <- clinical_theatre_data(clincial_example_df, time = date_yyyy_mm,
+#'                           name = theatre_name,
 #  wet_clinical_waste = clinical_waste_kg,
 #  wet_clinical_waste_unit = "kg",
 #  average = general_waste_kg,
