+++ DESCRIPTION
@@ -10,6 +10,8 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Suggests: 
+    knitr,
+    purrr,
     rmarkdown,
     testthat (>= 2.0.0)
 Config/testthat/edition: 2
@@ -22,12 +24,10 @@ Imports:
     ggplot2,
     ggpp,
     htmltools,
-    knitr,
     lubridate,
     magrittr,
     readxl,
     rlang,
-    purrr,
     shiny,
     shinydashboard,
     sp,
