### 2024-08-16 c67316e
Fixing notes - dontrun to examples which take >5 seconds
