+++ R/clinical_theatre_data.R
@@ -24,6 +24,7 @@
 #'
 #' @examples
 #' # Example with dummy data
+#' \dontrun{
 #' df <- data.frame(time = c("10/04/2000", "10/04/2000", "11/04/2000",
 #'                           "11/04/2000", "12/04/2000", "12/04/2000"),
 #' theatre = rep(c("A", "B"), times = 3),
@@ -31,7 +32,7 @@
 #' electricity_kwh = c(100, 110, 90, 100, 100, 110),
 #' general_waste = c(65, 55, 70, 50, 60, 30))
 #' 
-#' x <- clinical_theatre_data(df, time = time, name = theatre,
+#' clinical_theatre_data(df, time = time, name = theatre,
 #'                  wet_clinical_waste = clinical_waste,
 #'                  wet_clinical_waste_unit = "kg",
 #'                  desflurane = 10,
@@ -42,6 +43,7 @@
 #'                  jurisdiction = "Australia",
 #'                  year = 2023,
 #'                  single_sheet = FALSE)
+#' }
 clinical_theatre_data <- function(data, time, date_format = c("%d/%m/%Y"), name, wet_clinical_waste = 0, wet_clinical_waste_unit = c("tonnes", "kg"),
                                   desflurane = 0, sevoflurane = 0, isoflurane = 0, methoxyflurane = 0, N2O = 0, propofol = 0,
                                   water_supply = 0, water_trt = TRUE, water_unit = c("cubic metres", "million litres"),
