### 2024-08-28 #6932 @zhenchun
1. Line 15: Examples of the traditional tools?
2. Line 21: broader audience: please be specific.
3. Line 24: How did you connect the theatre emission to Hospital emission? I can not get the logic here. I think it needs more explanations.
4.  Line 30: This tool was based on the UK government guidelines. Is it transferable to other jurisdiction areas?
5. Line 47: Need more explanation on the theory behind the "Estimating Travel Emissions". Based on the fuel consumption divided by the number of passengers?
6. Nees a section of "Limitations"

We need a lot of improvement to clear the aims or the goals of this tool's application!
