### 2024-08-28 #6932 @mengqi-z
@zhenchun - Thank you for your comments on the paper. @lilyclements, could you please address Zhenchun's feedback on the paper, including his question about the warning message?

@zhenchun - Additionally, you may choose to open an issue on the `carbonr` GitHub repository [here](https://github.com/IDEMSInternational/carbonr/issues) for each problem you encounter during your review. Please tag the JOSS review link https://github.com/openjournals/joss-reviews/issues/6932 in the issue you open. This will automatically generate a message in the JOSS review thread, making it easier to track and resolve each issue.
