### 2024-08-29 37c8508
adding changes to paper after JOSS review
