+++ paper.md
@@ -27,26 +27,23 @@ bibliography: paper.bib
 ---
   
 # Summary
-The `carbonr` package provides a user-friendly, open-source tool for calculating carbon-equivalent emissions based on various sources, primarily the UK Government’s greenhouse gas reporting guidelines [@ukgov_greenhouse_gas_reporting_2023].
-
-Designed for flexibility and ease of use, `carbonr` enables users from different sectors to estimate emissions from travel, construction, office environments, and, more recently, clinical settings, thereby aiding in informed decision-making for environmental impact reduction.
+The `carbonr` package is an open-source, user-friendly tool for calculating carbon-equivalent emissions, primarily based on the UK Government’s greenhouse gas reporting guidelines [@ukgov_greenhouse_gas_reporting_2023]. 
+Designed for flexibility and ease of use, `carbonr` supports emissions estimation across various sectors, including travel, construction, office environments, and clinical settings, aiding informed environmental decision-making for environmental impact reduction.
 
 # Statement of Need
-Climate change and environmental sustainability are pressing global challenges that require sophisticated methods for assessing and mitigating carbon emissions across various sectors [@WHO2021]. Accurate carbon accounting is important for understanding and reducing environmental impacts; however, the tools available to researchers and practitioners often fall short in terms of flexibility, transparency, and accessibility.
+Addressing climate change and environmental sustainability requires advanced methods for assessing and mitigating carbon emissions [@WHO2021]. Many existing tools available to researchers and practitioners, such as the UK Government's Excel templates [@ukgov_greenhouse_gas_reporting_2023] and online calculators [@carbonfootprint_calculator], often involve manual calculations using spreadsheets or proprietary software.
+These tools, while comprehensive, are often not easily adaptable to different sectors and can lack the reproducibility and automation of open-source solutions. These limitations can hamper scientific advancements and the widespread adoption of best practices in accounting for carbon emissions. 
+Open-source tools like co2calculator [@pledge4future_co2calculator] and footprint [@rpackage_footprint] cover limited areas, such as specialising in air travel, and do not meet the diverse needs of users who require a broader range of emission activities.
 
-Traditional carbon accounting tools often involve manual calculations using spreadsheets or proprietary software. For example, the UK Government provides Excel-based templates {@ukgov_greenhouse_gas_reporting_2023}, and the Carbon Footprint Calculator is one online calculator to estimate emissions [@carbonfootprint_calculator]. These tools, while comprehensive, are often not easily adaptable to different sectors or use-cases. Additionally, they lack the reproducibility offered by an open-source solution which could automate calculations and provide a more streamlined, adaptable approach. These limitations can hamper scientific advancements and the widespread adoption of best practices in accounting for carbon emissions. 
+`carbonr` addresses these gaps by offering a flexible, open-source solution that improves the reproducibility of carbon emission calculations and enhances the understanding of these emissions within the research community.
 
-Current open-source solutions have a limited scope; for example, `co2calculator` in Python focusses on calculations related to heating, electricity, business trips, and transportation [@pledge4future_co2calculator]. Similarly, `footprint` in R specialises in air travel emissions [@rpackage_footprint]. These tools are not comprehensive enough to address the diverse needs of users who require a broader range of emission activities.
+One key feature of `carbonr` is its user-friendly Shiny interface, broadening access to include non-technical users to engage in carbon management.
 
-To address these issues, `carbonr` offers an open-source solution that improves the reproducibility of carbon emission calculations and enhances the understanding of these emissions within the research community. Key features include its simplicity, user-friendly Shiny interface, and adaptability to different sector needs, making it accessible to a broader audience.
[…]
