+++ paper.bib
@@ -46,3 +46,14 @@
   year = {2021},
   url = {https://cran.r-project.org/web/packages/footprint/index.html}
 }
+
+@article{macneill2017impact,
+  title={The impact of surgery on global climate: a carbon footprinting study of operating theatres in three health systems},
+  author={MacNeill, Andrea J. and Lillywhite, Robert and Brown, Chris J.},
+  journal={The Lancet Planetary Health},
+  volume={1},
+  number={9},
+  pages={e381--e388},
+  year={2017},
+  publisher={Elsevier}
+}
