### 2024-08-29 #6932 @lilyclements
@zhenchun thank you. I have addressed the issues [here](https://github.com/IDEMSInternational/carbonr/pull/38)

I will get to the warning message shortly. Thank you.
