+++ tests/testthat/test-airplane_emissions.R
@@ -1,4 +1,3 @@
-library(airportr)
 library(carbonr)
 
 # load airport data
