### 2024-08-29 #6932 @lilyclements
@zhenchun the warning message has been fixed [here](https://github.com/IDEMSInternational/carbonr/pull/39). This is ready for a re-review. Thanks!
