### 2024-09-09 IDEMSInternational/carbonr#41 @hvwaldow
JOSS-review : "Via" - parameters in "Ferry" by shiny_emissions are not accepted.
This is the same symptom as in Issue #26 (which was fixed) , just with "Ferry" now instead of "Rail".
