### 2024-09-09 #6932 @hvwaldow
Hi @lilyclements, I just failed to get the"Via"-Option for Ferry in the Shiny frontend to work. This looks very similar to https://github.com/IDEMSInternational/carbonr/issues/26, which you already fixed. The new issue is https://github.com/IDEMSInternational/carbonr/issues/41.
