### 2024-09-09 a557429
fixing bug in shiny_emissions via parameter
