### 2024-09-09 IDEMSInternational/carbonr#43 @hvwaldow
JOSS-review : Incorrect Reference
I don't find anything called "Team" in the reference "Team, R.C., 2021" Author names are different. Also the year seems to be incorrect.
