+++ paper.bib
@@ -41,7 +41,7 @@
 }
 
 @misc{rpackage_footprint,
-  author = {R Core Team},
+  author = {Schmidt, Anthony and Kulma, Kasia},
   title = {footprint: Carbon Footprint Calculator for R},
   year = {2021},
   url = {https://cran.r-project.org/web/packages/footprint/index.html}
