+++ paper.bib
@@ -43,7 +43,7 @@
 @misc{rpackage_footprint,
   author = {Schmidt, Anthony and Kulma, Kasia},
   title = {footprint: Carbon Footprint Calculator for R},
-  year = {2021},
+  year = {2024},
   url = {https://cran.r-project.org/web/packages/footprint/index.html}
 }
 
