### 2024-09-09 IDEMSInternational/carbonr#43 @lilyclements
@hvwaldow thanks so much for spotting this - you're absolutely right on the reference which I've just fixed. Let me know if you're happy so I can close this issue :)
