### 2024-09-20 IDEMSInternational/carbonr#45 @mengqi-z
JOSS Paper
Hi @lilyclements,

I have a few comments on the paper that need to be addressed before we proceed to the next stage:

1. Lines 52-53 duplicate the content from lines 49-51. I recommend removing lines 52-53 or merging with previous paragraph to avoid redundancy.

2. Usage Examples. On lines 62 and 67, please prepend `carbonr::` to the functions to ensure consistency with the formatting used for other functions in the examples.

3. References.

   - For MacNeil et al., 2017.  Please add the DOI: `10.1016/S2542-5196(17)30162-6`
   - For Schmidt and Kulma, 2024. Please update the link to the appropriate DOI: `10.32614/CRAN.package.footprint`

Thank you.

This issue is for JOSS paper review https://github.com/openjournals/joss-reviews/issues/6932
