### 2024-09-20 #6932 @mengqi-z
Hi @lilyclements - I’ve proofread the paper and created an issue for a few minor improvements. Could you please review and address these, and let me know once you’ve completed them? Thank you!
