### 2024-09-21 958f9fb
addressing issue #45
