+++ paper.bib
@@ -44,7 +44,7 @@
   author = {Schmidt, Anthony and Kulma, Kasia},
   title = {footprint: Carbon Footprint Calculator for R},
   year = {2024},
-  url = {https://cran.r-project.org/web/packages/footprint/index.html}
+  doi = {10.32614/CRAN.package.footprint}
 }
 
 @article{macneill2017impact,
@@ -55,5 +55,6 @@
   number={9},
   pages={e381--e388},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/S2542-5196(17)30162-6}
 }
