### 2024-09-21 #6932 @lilyclements
@mengqi-z thank you, I have addressed these in [PR 46](https://github.com/IDEMSInternational/carbonr/issues/46) on my branch and merged them in
