### 2024-10-01 #6932 @lilyclements
@mengqi-z I have done the tasks you outlined. The version is **v.0.2.0** and the DOI is **10.5281/zenodo.13871038**

[x]  Double check authors and affiliations (including ORCIDs).
[x]  Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper. - This is version **0.2.0**
[x]  Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
[x]  Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
[x]  Make sure that the license listed for the archive is the same as the software license.
[x]  Please **post the DOI of the archived version here**: **10.5281/zenodo.13871038**
