### 2024-10-01 #6932 @mengqi-z
Thanks @lilyclements .

I noticed that there is a discrepancy between the licenses listed on Zenodo and GitHub. The Zenodo archive currently specifies the `GNU General Public License v3.0 or later`, while the GitHub repository uses the `GNU Lesser General Public License version 3` (LGPL v3).

To ensure consistency, could you please update the license on Zenodo to match the LGPL v3 license used on GitHub? I have confirmed that the LGPL v3 license is available on Zenodo. Once updated, please share the latest DOI here.
