### 2024-10-01 #6932 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
