+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: carbonr
 Title: Calculate CO2e emissions
-Version: 0.2.0
+Version: 0.2.1
 Authors@R: 
     person("Lily", "Clements", , "lily@idems.international", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-8864-0552"))
