### 2024-10-08 e6cc85f
Adding vignettes on personal use
