+++ vignettes/Personal_carbon_emissions.Rmd
@@ -1,0 +1,245 @@
+---
+title: "Introduction to carbonr"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Personal Carbon Emissions}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+## Personal Carbon Emissions with carbonr
+
+### Introduction
+Personal carbon emissions refer to the total amount of greenhouse gases, primarily carbon dioxide (CO₂), that an individual is responsible for emitting directly and indirectly through their daily activities.The carbonr package in R provides a convenient way to calculate these emissions.
+The emissions values used in these calculations primarily derive from the [UK Government report (2023)](https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2023), where available. If calculations are based on sources other than the UK Government report, it will be specified accordingly.
+
+### Motivation
+The aim of calculating Personal Carbon Emissions is to provide a reliable and reproducible method for assessing individual carbon footprints. This ensures that results can be easily saved, edited, and redistributed, fostering transparency and accuracy in carbon emission reporting.
+
+### Installation
+To install the development version of carbonr, you can use the following commands to download it from [GitHub](https://github.com/):
+
+```{r install_package, eval = FALSE}
+# Uncomment the following line if you haven't installed devtools yet
+# install.packages("devtools")
+devtools::install_github("IDEMSInternational/carbonr")
+```
+
+Once installed, load the carbonr package with the following command:
+```{r, eval=FALSE}
+library(carbonr)
+library(knitr)
+library(rmarkdown)
+```
+
+### Usage
+The carbonr package allows users to calculate personal emissions from various sources, including travel-related activities, building energy use, and household emissions. For personal emissions, we consider the following categories:
+```
+*Travel-Related Emissions*:
+
+airplane_emissions() Calculates emissions from air travel.
+ferry_emissions() Estimates emissions from ferry travel
+land_emissions() Computes emissions from land-based travel.
+rail_emissions() Determines emissions from rail travel.
+vehicle_emissions() Assesses emissions from personal vehicle use.
+
+*Household Emissions*:
+household_emissions() Estimates emissions from household activities.
+
+*Building Emissions*:
+building_emissions() Calculates emissions associated with building energy use.
+```
+
+These all return carbon-equivalent emissions in tonnes.
+
+A shiny app is also available by shiny_emissions() to calculate carbon-equivalent emissions with a GUI.
+
+We give some small examples in using the functions in `carbonr()`.
+We started by Travel related Emissions:
+1.`airplane_emissions()`
+To calculate emissions for a flight between Nairobi and Kisumu, we first want to find the name of the airports. We do this using the `airport_finder()` function:
+
+```{r load_tidyverse, message = FALSE, warning = FALSE}
+library(tidyverse)
+```
+
+
+```{r find_airport_nairobi, message = FALSE, warning = FALSE, eval=FALSE, include=TRUE}
+airport_finder(name = "Nairobi")
+ %>%
+  knitr::kable()
+```
+
+
+
+```{r get_airport_kisumu, message = FALSE, warning = FALSE, eval=FALSE, include=TRUE}
+airport_finder(name = "Kisumu")
+ %>%
+  knitr::kable()
+
+```
+
+Now we can find the overall emission value using the appropriate IATA code. These distances are calculated using the Haversine formula:
+
[…]
