### 2024-10-09 3c1b3d9
update URLs and Descriptions
