+++ R/stations.R
@@ -13,6 +13,6 @@
 #'   \item{latitude}{Latitude of the station}
 #'   \item{longitude}{Longitude of the station}
 #' }
-#' @source \url{https://data.gov.uk/dataset/ff93ffc1-6656-47d8-9155-85ea0b8f2251/national-public-transport-access-nodes-naptan}
+#' @source \url{https://www.data.gov.uk/dataset/ff93ffc1-6656-47d8-9155-85ea0b8f2251/national-public-transport-access-nodes-naptan}
 #' @source \url{https://www.theguardian.com/news/datablog/2011/may/19/train-stations-listed-rail#data}
 "stations"
