+++ man/total_output.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/total_output_plot.R
+% Please edit documentation in R/total_output.R
 \name{total_output}
 \alias{total_output}
 \title{Calculate Total Output and Generate Plot}
@@ -38,11 +38,3 @@ This function calculates the total output by grouping the data based on the spec
 The resulting data is used to create a line plot showing the total output over time, with each group represented
 by a different color. The plot can be grouped by the default grouping, month, or year, based on the plot_by parameter.
 }
-\examples{
-# todo
-
-}
-\keyword{grouping}
-\keyword{output,}
-\keyword{plot,}
-\keyword{total}
