+++ man/airports.Rd
@@ -28,8 +28,6 @@ A data frame with 7698 rows and 14 variables:
 }
 }
 \source{
-\url{https://openflights.org/data.html}
-
 \url{https://cran.r-project.org/package=airportr}
 }
 \usage{
