+++ README.Rmd
@@ -132,7 +132,7 @@ Your feedback and contributions are invaluable in helping us improve and maintai
 ## References
 
 #### Other Online Calculators:
-- [Carbonfund.org](https://carbonfund.org/calculation-methods/)
+- [Carbonfund.org](https://shop.climeco.com/individual/individual-emissions-calculator/)
 - [Carbon Footprint Calculator](https://www.carbonfootprint.com/calculatorfaqs.html)
 
 #### Sources
