+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: carbonr
-Title: Calculate carbon-equivalent emissions.
+Title: Calculate Carbon-Equivalent Emissions
 Version: 0.2.1
 Authors@R: 
     person("Lily", "Clements", , "lily@idems.international", role = c("aut", "cre"),
